predictors/iu/requirements_event_place.txt
absl-py==0.8.1
astor==0.8.0
astunparse==1.6.3
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cloudpickle==1.1.1
cmake==3.20.2
cycler==0.10.0
decorator==4.4.1
dlib==19.21.0
gast==0.3.3
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
imutils==0.5.3
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
llvmlite==0.31.0
Markdown==3.1.1
matplotlib==3.1.1
networkx==2.4
numba==0.48.0
numpy==1.17.4
oauthlib==3.1.0
opencv-contrib-python==4.1.1.26
opencv-python==3.4.2.17
opt-einsum==3.1.0
pandas==1.1.1
Pillow==6.2.1
progressbar2==3.47.0
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyparsing==2.4.5
python-dateutil==2.8.1
python-utils==2.3.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.4.1
six==1.13.0
tensorboard==2.3.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
tensorflow-probability==0.8.0rc0
termcolor==1.1.0
tf-slim==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
